clase3/requirements.txt
flask
python-dotenv
flask-restful
flask_sqlalchemy